[REFACTOR] 디렉토리 구조 변경

+++ src/main/java/com/project/trysketch/suggest/AdjectiveRepository.java
@@ -3,5 +3,5 @@ package com.project.trysketch.suggest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 // 1. 기능   : 형용사 Repository
-public interface AdjectiveRepository extends JpaRepository<AdjectiveEntity, Long> {
+public interface AdjectiveRepository extends JpaRepository<AdjectiveEntity, Integer> {
 }
